requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Base de datos
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validación y configuración
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
bcrypt>=4.0.1
python-multipart==0.0.9

# WebSockets (modo online)
websockets==12.0

# HTTP client (para llamadas a IA externa)
httpx>=0.28.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6

# Archivos estáticos (frontend)
aiofiles==23.2.1

# Utilidades
python-dateutil==2.9.0
resend==2.26.0